refactor: simplify MatchHistoryPanel by extracting MatchHistoryTable component
